- finish implementing permissions and check their functionality

+++ lib/modules/appointments/screens/appointment-details-map.dart
@@ -267,7 +267,7 @@ class AppointmentDetailsMapState extends State<AppointmentDetailsMap>
     Provider.of<WorkEstimateProvider>(context).defaultQuantity = (_provider
                 .selectedAppointmentDetail.auctionMapDirections.totalDistance /
             1000)
-        .round();
+        .round() as double;
     Provider.of<WorkEstimateProvider>(context).maxDate = _provider
         .selectedAppointmentDetail
         .auctionMapDirections
